Move scroll button file
